backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
chromadb
langchain
langchain-openai
langchain-community
pymupdf
openai